V5
I added image upload for background, and I added a way to upload mp3 for custom sounds when you hit rate chat (it only makes a sound after you hit ok after reading the sus level, and the app will look like it froze, but it is just a delay). New button layout.